pyproject.toml
[project]
name = "coffee-shop-data-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
]
